requirements.txt
html2text==2020.1.16
PyYAML==5.4.1
requests==2.25.1
pycron==3.0.0
